+++ agro360-dto/src/main/java/com/agro360/dto/common/AbstractLigneDto.java
@@ -1,20 +1,21 @@
 package com.agro360.dto.common;
 
-import javax.persistence.Column;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.MappedSuperclass;
+import java.math.BigDecimal;
 
 import com.agro360.dto.stock.ArticleDto;
 import com.agro360.dto.stock.UniteDto;
 import com.agro360.dto.stock.VariantDto;
 import com.agro360.vd.stock.TypeLigneEnumVd;
 
+import jakarta.persistence.Column;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.MappedSuperclass;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 
@@ -54,6 +55,6 @@ public abstract class AbstractLigneDto extends AbstractDto {
 	private Double quantite;
 
 	@Column(name = "PRIX_UNITAIRE", nullable = false, precision = 16, scale = 4)
-	private Double prixUnitaire;
+	private BigDecimal prixUnitaire;
 
 }
